+++ source/Api/EventListeners/EventDetails.cs
@@ -12,17 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Microsoft.Azure.Functions.Worker;
+namespace Api.EventListeners;
 
-namespace Api.Configuration.IntegrationEvents;
-
-internal static class FunctionContextExtensions
+public record EventDetails(string EventId, string EventType)
 {
-    internal static EventDetails ExtractEventDetails(this FunctionContext context)
+    public static EventDetails Empty()
     {
-        context.BindingContext.BindingData.TryGetValue("MessageId", out var messageId);
-        context.BindingContext.BindingData.TryGetValue("Label", out var subject);
-
-        return messageId is string eventId && subject is string eventType ? new EventDetails(eventId, eventType) : EventDetails.Empty();
+        return new EventDetails(string.Empty, string.Empty);
     }
 }
